Edit user

+++ ChattingiOS/Main/Flow.swift
@@ -12,14 +12,21 @@ final class Flow {
     private var contentViewModel: ContentViewModel { dependencies.contentViewModel }
     private var currentUserVault: CurrentUserVault { dependencies.currentUserVault }
     private var pushNotificationHandler: PushNotificationsHandler { dependencies.pushNotificationHandler }
-    private var navigationControl: NavigationControlViewModel { contentViewModel.navigationControl }
     private var style: ViewStyleManager { dependencies.viewStyleManager }
+    
+    private var navigationControlForContacts: NavigationControlViewModel {
+        contentViewModel.navigationControlForContacts
+    }
+    private var navigationControlForProfile: NavigationControlViewModel {
+        contentViewModel.navigationControlForProfile
+    }
     
     // Let Flow manage the lifetime of the ContactListViewModel instance. Since there's a weird behaviour,
     // the ContactListView sometime will not update if let it manage its own ContactListViewModel.
     private var contactListViewModel: ContactListViewModel?
     
     private weak var messageListViewModel: MessageListViewModel?
+    private var editProfileTask: Task<Void, Never>?
     private var newContactTask: Task<Void, Never>?
     var deviceToken: String? {
         didSet {
@@ -47,8 +54,8 @@ final class Flow {
             
             guard let user = await currentUserVault.retrieveCurrentUser()?.user else { return }
             
-            await contentViewModel.set(signInState: .signedIn(user))
-            navigationControl.forceReloadContent()
+            await contentViewModel.set(signInState: .signedIn(user, to: .contacts))
+            navigationControlForContacts.forceReloadContent()
         }
     }
     
@@ -67,9 +74,9 @@ final class Flow {
         defer { contentViewModel.isLoading = false }
         
         // Order does matter!
-        await contentViewModel.set(signInState: .signedIn(user))
+        await contentViewModel.set(signInState: .signedIn(user, to: .contacts))
         contactListViewModel = nil
-        navigationControl.forceReloadContent()
+        navigationControlForContacts.forceReloadContent()
         await updateDeviceToken()
     }
     
@@ -93,7 +100,7 @@ final class Flow {
             guard let currentUserID = contentViewModel.user?.id, currentUserID == userID else { return }
             
             // On contacts tab, not on MessageListView.
-            if contentViewModel.selectedTab == .contacts, navigationControl.path.count < 1 {
+            if contentViewModel.selectedTab == .contacts, navigationControlForContacts.path.count < 1 {
                 showMessageListView(currentUserID: currentUserID, contact: contact)
             }
             
@@ -128,21 +135,18 @@ final class Flow {
     func startView() -> some View {
         ContentView(viewModel: contentViewModel) { [unowned self] currentUser in
             TabView(selection: contentViewModel.selectedTabBinding) { [unowned self] in
-                NavigationControlView(viewModel: navigationControl) { [unowned self] in
+                NavigationControlView(viewModel: navigationControlForContacts) { [unowned self] in
                     contactListView(currentUserID: currentUser.id)
                 }
-                .tabItem {
-                    Label(TabItem.contacts.title, systemImage: TabItem.contacts.systemImage)
-                }
+                .tabItem { Label(TabItem.contacts.title, systemImage: TabItem.contacts.systemImage) }
                 .tag(TabItem.contacts)
                 
-                profileView(user: currentUser)
-                    .tabItem {
-                        Label(TabItem.profile.title, systemImage: TabItem.profile.systemImage)
-                    }
-                    .tag(TabItem.profile)
+                NavigationControlView(viewModel: navigationControlForProfile) { [unowned self] in
+                    profileView(user: currentUser)
+                }
+                .tabItem { Label(TabItem.profile.title, systemImage: TabItem.profile.systemImage) }
+                .tag(TabItem.profile)
             }
-            .toolbarColorScheme(.dark, for: .tabBar)
         } signInContent: { [unowned self] in
             signInView()
         } sheet: { [unowned self] in
@@ -178,14 +182,62 @@ final class Flow {
         return SignUpView(viewModel: viewModel)
     }
     
-    private func profileView(user: User) -> ProfileView {
+    private func profileView(user: User) -> some View {
         let viewModel = ProfileViewModel(user: user, loadImageData: dependencies.decoratedLoadImageDataWithCache)
-        return ProfileView(viewModel: viewModel, signOutAction: { [unowned self] in
-            Task {
-                try? await currentUserVault.deleteCurrentUser()
-                await contentViewModel.set(signInState: .userInitiatedSignOut)
+        return ProfileView(
+            viewModel: viewModel,
+            editAction: { [unowned self] avatarData in
+                showEditProfileView(user: user, avatarData: avatarData)
+            },
+            signOutAction: { [unowned self] in
+                Task {
+                    try? await currentUserVault.deleteCurrentUser()
+                    await contentViewModel.set(signInState: .userInitiatedSignOut)
+                }
             }
+        )
+        .navigationDestinationFor(EditProfileView.self)
+    }
+    
+    private func showEditProfileView(user: User, avatarData: Data?) {
+        let viewModel = EditProfileViewModel(
+            user: user,
+            currentAvatarData: avatarData,
+            updateCurrentUser: dependencies.updateCurrentUser
+        )
+        editProfileTask?.cancel()
+        editProfileTask = Task { [unowned self] in
+            defer { editProfileTask = nil }
+            
+            for await editedUser in viewModel.$user.values {
+                guard editedUser != user else { continue }
+                
+                if let currentUser = await currentUserVault.retrieveCurrentUser() {
+                    do {
+                        // Save edited user into current user vault.
+                        try await currentUserVault.saveCurrentUserWithoutNotifyObservers(
+                            user: editedUser,
+                            token: Token(accessToken: currentUser.accessToken, refreshToken: currentUser.refreshToken)
+                        )
+                        
+                        await contentViewModel.set(signInState: .signedIn(editedUser, to: .profile))
+                        navigationControlForProfile.forceReloadContent()
+                    } catch {
+                        // If save error occurred, delete the current user, force user sign in.
+                        try? await currentUserVault.deleteCurrentUser()
+                    }
+                    
+                    // Release this task after user update process.
+                    editProfileTask?.cancel()
+                } else {
+                    assertionFailure("CurrentUser should not be nil just after edit profile.")
+                }
+            }
+        }
+        let editProfileView = EditProfileView(viewModel: viewModel, onDisappear: { [unowned self] in
+            editProfileTask?.cancel()
         })
+        navigationControlForProfile.show(next: NavigationDestination(editProfileView))
     }
     
     private func contactListView(currentUserID: Int) -> some View {
@@ -246,6 +298,6 @@ final class Flow {
         messageListViewModel = viewModel
         
         let destination = NavigationDestination(MessageListView(viewModel: viewModel))
-        navigationControl.show(next: destination)
+        navigationControlForContacts.show(next: destination)
     }
 }
